+++ test/send.js
@@ -483,6 +483,18 @@ describe('send(file, options)', function(){
       .expect('Cache-Control', 'public, max-age=123', done)
     })
 
+    it('should accept string', function(done){
+      var app = http.createServer(function(req, res){
+        send(req, 'test/fixtures/name.txt')
+        .maxage('1y')
+        .pipe(res);
+      });
+
+      request(app)
+      .get('/name.txt')
+      .expect('Cache-Control', 'public, max-age=31557600', done)
+    })
+
     it('should support Infinity', function(done){
       var app = http.createServer(function(req, res){
         send(req, 'test/fixtures/name.txt')