+++ History.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+ * Accept string for `maxAge` (converted by `ms`)
  * Include link in default redirect response
 
 0.4.3 / 2014-06-11